Add PCI Ids and fix hash calculation
Added the PCI detection needed for device bringup. Only strange thing is that it
looks like 9377 uses 6174 registers. Not sure why.

Initially a page fault was caused by SKB_RX_HASH generating an index longer than
the hash table. I changed it to use a modulus instead of the ~(32 - 1). I'm not
sure why this original value was chosen, or why it causes a fault.